chore(api): migrate target-profile-summary-for-admin endpoint and controller
