+++ api/lib/application/target-profiles/index.js
@@ -1,54 +1,11 @@
 import Joi from 'joi';
 
-import { BadRequestError, sendJsonApiError } from '../../../src/shared/application/http-errors.js';
 import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
 import { targetProfileController } from './target-profile-controller.js';
 
 const register = async function (server) {
   server.route([
-    {
-      method: 'GET',
-      path: '/api/admin/target-profile-summaries',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        validate: {
-          options: {
-            allowUnknown: true,
-          },
-          query: Joi.object({
-            filter: Joi.object({
-              id: Joi.number().integer().empty('').allow(null).optional(),
-              name: Joi.string().empty('').allow(null).optional(),
-            }).default({}),
-            page: Joi.object({
-              number: Joi.number().integer().empty('').allow(null).optional(),
-              size: Joi.number().integer().empty('').allow(null).optional(),
-            }).default(),
-          }),
-          failAction: (request, h) => {
-            return sendJsonApiError(new BadRequestError('Un des champs de recherche saisis est invalide.'), h);
-          },
-        },
-        handler: targetProfileController.findPaginatedFilteredTargetProfileSummariesForAdmin,
-        tags: ['api', 'admin', 'target-profiles'],
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            '- Elle permet de récupérer & chercher une liste de profils cible\n' +
-            '- Cette liste est paginée et filtrée selon un **id** et/ou un **name** donnés',
-        ],
-      },
-    },
     {
       method: 'GET',
       path: '/api/admin/target-profiles/{id}',
