+++ api/lib/application/target-profiles/target-profile-controller.js
@@ -1,20 +1,9 @@
 import { usecases as devcompUsecases } from '../../../src/devcomp/domain/usecases/index.js';
 import * as trainingSummarySerializer from '../../../src/devcomp/infrastructure/serializers/jsonapi/training-summary-serializer.js';
 import * as targetProfileSerializer from '../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-serializer.js';
-import * as targetProfileSummaryForAdminSerializer from '../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-summary-for-admin-serializer.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import * as targetProfileForAdminSerializer from '../../infrastructure/serializers/jsonapi/target-profile-for-admin-serializer.js';
-
-const findPaginatedFilteredTargetProfileSummariesForAdmin = async function (request) {
-  const options = request.query;
-
-  const { models: targetProfileSummaries, meta } = await usecases.findPaginatedFilteredTargetProfileSummariesForAdmin({
-    filter: options.filter,
-    page: options.page,
-  });
-  return targetProfileSummaryForAdminSerializer.serialize(targetProfileSummaries, meta);
-};
 
 const getTargetProfileForAdmin = async function (request, h, dependencies = { targetProfileForAdminSerializer }) {
   const targetProfileId = request.params.id;
@@ -61,7 +50,6 @@ const findPaginatedTrainings = async function (request, h, dependencies = { trai
 };
 
 const targetProfileController = {
-  findPaginatedFilteredTargetProfileSummariesForAdmin,
   getTargetProfileForAdmin,
   updateTargetProfile,
   createTargetProfile,
